update generate.py
When generating lyrics, a random word is chosen based on the model's probability distribution for the next word.  This is a change from simply picking the word with the highest probability.  Making this change allows for different lyrics to be generated based on the model when providing the same input sequence.